requirements.txt
asgiref==3.5.0
Django==4.1.4
sqlparse==0.4.2
psycopg2==2.9.5
djangorestframework==3.14.0
markdown
django-filter
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
Pillow==23.0.1
django-environ==0.10.0